Add new port net/cagibi:
Cagibi is an experimental cache/proxy system for the SSDP* (Simple
Service Discovery Protocol) part of UPnP.

Cagibi aims to be to SSDP what Avahi is to DNS-SD/Zeroconf: a cache
caching all service/device announcements on the network in a local
process as well as being a broker serving local announcements to
the network.  Both should be done by a single daemon process,
accessable via D-Bus on the system bus.  The cache should offer
active queries, so another process is only informed about changes
about UPnP devices it is interested in.