Add the missing FreeBSD SA entries. We used to add these but stopped a while back. This should catch us up.
According to cperciva@ the reason we stopped was that it was causing a lot of
false positives. I ran portaudit with these changes and did not see any false
positives but if it turns out to be too noisy I will remove them.

Submitted by:	Christopher J. Umina (private mail)
Approved by:	cperciva@