Rework the previous fix to sync. with upstream:
http://hg.openjdk.java.net/jdk7/jdk7/jdk/diff/441f88d39988/src/solaris/native/java/net/net_util_md.c

Note it has no functional change for FreeBSD because sizeof(socklen_t) is
luckily sizeof(int).  Actually, this commit is just to note that the regression
was acidentally pulled in from upstream because of my ignorance.  Specifically,
the following commit was the culprit:

http://hg.openjdk.java.net/jdk7/jdk7/jdk/diff/b7fce4bac617/src/solaris/native/java/net/net_util_md.c

FYI, it also triggered the following upstream PR:

http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6730740

Obtained from:	OpenJDK project
Pointyhat:	jkim