http://www.linuxtv.org/vdrwiki/index.php/Control-plugin
The 'control' plugin brings the ability to VDR to control
the whole OSD over a telnet client.

To reach this, 'control' listens on a network socket
(default is port 2002). If a client wants to connect, VDR
checks if that client is allowed to connect to VDR (see in
the documentation of VDR about the svdrphosts.conf file for
more info). If the connection is etablished, 'control'
sends the curent OSD state to the client. Also all key
strokes at the client side are redirected to VDR.

WWW: http://ricomp.de/vdr/down_en.html