Don't rebuild manual pages if configure was changed
Sometimes patches for PostgreSQL touch its configure script.  And the
logics inside current Makefiles for the documentation makes the manual
pages (and other stuff) to be dependent of the timestamp of the
top-level configure file.  This triggers the rebuild of the manual
pages, but since some additional XML-related tools are needed, this
rebuild can fail.

PR: 159844
Approved by: maintainer timeout (1 month)
Feature safe: yes