Welcome the new lang/gcc port!  This shall track our preferred version of GCC (usually an upstream release).  It starts out as GCC 4.6.1 and is thus in conflict with lang/gcc46 and will move towards later minor versions of GCC 4.6 and then on to GCC 4.7.
lang/gcc will provide gcc46, g++46, gfortran46 etc. exactly like
lang/gcc46 with which it is interchangible.

This is also planned to be in sync with our existing USE_FORTRAN knob
so that users have the option of using this port, rarely updated, or
the corresponding lang/gcc46 which follows weekly upstream snapshots.

On the way rename %%GCC_VER%% in pkg-plist to %%GCC_VERSION%% and
make the Makefile machinery a bit more generic to minimize differences
between lang/gcc ports based on releases and those based on snapshots.

PR:		156857